Moved Delete Everything to Settings
QoL changes, moved delete all to settings, "List" now renamed to "Medicaments"
