+++ .idea/misc.xml
@@ -13,9 +13,11 @@
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/content_main.xml" value="0.35104166666666664" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_add.xml" value="0.33" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_list.xml" value="0.33" />
+        <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_settings.xml" value="0.35104166666666664" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_update.xml" value="0.20969945355191258" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/nav_header_main.xml" value="0.35104166666666664" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/text_row_item.xml" value="0.33" />
+        <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/menu/activity_main_drawer.xml" value="0.35104166666666664" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/menu/delete_menu.xml" value="0.33" />
         <entry key="..\:/Users/z0045evs/Desktop/AndroidDev/healthnote-android/app/src/main/res/drawable/ic_menu_list.xml" value="0.2315" />
         <entry key="..\:/Users/z0045evs/Desktop/AndroidDev/healthnote-android/app/src/main/res/layout/activity_add_medicament.xml" value="0.25" />
