+++ app/src/main/java/de/othregensburg/healthnote/fragments/list/ListFragment.kt
@@ -1,9 +1,9 @@
 package de.othregensburg.healthnote.fragments.list
 
-import android.app.AlertDialog
 import android.os.Bundle
-import android.view.*
-import android.widget.Toast
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -66,30 +66,5 @@ class ListFragment : Fragment() {
 
         return view
     }
-
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.delete_menu, menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == R.id.menu_delete) {
-            deleteAllMeds()
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    private fun deleteAllMeds() {
-        val builder = AlertDialog.Builder(requireContext())
-        builder.setPositiveButton("Yes") {_, _ ->
-            mMedViewModel.deleteAllMeds()
-            Toast.makeText(requireContext(), "Successfully removed everything", Toast.LENGTH_SHORT).show()
-        }
-        builder.setNegativeButton("No") {_, _ -> }
-        builder.setTitle("Delete everything?")
-        builder.setMessage("Are you sure you want to delete everything?")
-        builder.create().show()
-    }
-
 
 }
