+++ app/src/main/res/layout/fragment_settings.xml
@@ -39,4 +39,16 @@
         android:layout_marginBottom="8dp"
         android:text="@string/enableBiometrics"
         android:textSize="16sp"/>
+
+    <TextView
+        android:id="@+id/deleteMedicaments"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="8dp"
+        android:text="@string/deleteAllMeds"
+        android:textSize="16sp"
+        android:textColor="@color/critical"/>
 </LinearLayout>
