+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -6,7 +6,7 @@
     <group android:checkableBehavior="single">
         <item
             android:id="@+id/listFragment"
-            android:icon="@drawable/ic_menu_list"
+            android:icon="@drawable/ic_baseline_medical_services_24"
             android:title="@string/menu_list" />
         <item
             android:id="@+id/calendarFragment"
